Add transactions

+++ core/History.hpp
@@ -17,8 +17,8 @@ public:
     Q_PROPERTY(QString target READ getTarget)
     QString target;
 
-    Q_PROPERTY(int value READ getValue)
-    int value;
+    Q_PROPERTY(double value READ getValue)
+    double value;
 
     Q_PROPERTY(QString date READ getDate)
     QString date;
@@ -31,7 +31,7 @@ public:
     QString getTarget() const { return target; };
 
     [[nodiscard]]
-    int getValue() const { return value; }
+    double getValue() const { return value; }
 
     [[nodiscard]]
     QString getDate() const { return date; }
